==> analysis/01_data_processing.Rmd <==
---
title: "ManyPrimates_Pilot_Data_Processing"
author: "Manuel Bohn"
date: "7 Oct 2018"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r, message=FALSE}
library(tidyverse)
library(readxl)
```

```{r sample, message=FALSE}
# load sample
path <- "../data/raw_data/"
sample <- read_delim(str_c(path, "coding_sample.csv"), delim = ";")
```

```{r}
# define helper function
fail_str <- "--- FAIL ---\n"

do_checks <- function(df) {
  cat("check column number (= 19?): ")
  if (ncol(sample) == ncol(df)) cat("pass\n")
  else {cat(fail_str); print(ncol(df))}

  cat("\ncheck column names: ")
  if (all(names(sample) %in% names(df))) cat("pass\n")
  else {cat(fail_str); print(names(df))}

  cat("\ncheck condition names: ")
  if (all(unique(df$condition) %in% unique(sample$condition))) cat("pass\n")
  else {cat(fail_str); print(unique(df$condition))}

  cat("\ncheck species:\n")
  print(unique(df$species))

  cat("\ncheck subject:\n")
  print(sort(unique(df$subject_site), na.last = T))

  cat("\ncheck age:\n")
  print(sort(unique(df$age), na.last = T))

  cat("\ncheck cup distance:", unique(df$cup_distance), "\n")

  cat("\ncheck task experience: ")
  if (all(unique(df$task_experience) %in% unique(sample$task_experience))) cat("pass\n")
  else {cat(fail_str); print(unique(df$task_experience))}

  cat("\ncheck board size:", unique(df$board_size), "\n")

  cat("\ncheck test situation: ")
  if (all(unique(df$test_situation) %in% unique(sample$test_situation))) cat("pass\n")
  else {cat(fail_str); print(unique(df$test_situation))}

  cat("\ncheck session:\n")
  print(unique(df$session))

  cat("\ncheck block:\n")
  print(unique(df$block))

  cat("\ncheck trial:\n")
  print(sort(unique(df$trial), na.last = T))

  cat("\ncheck pick:", sort(unique(df$pick), na.last = T), "\n")

  cat("\ncheck hiding locations: ")
  if (all(unique(df$hiding_location) %in% unique(sample$hiding_location))) cat("pass\n")
  else {cat(fail_str); print(unique(df$hiding_location))}

  cat("\ncheck correct:", unique(df$correct), "\n")
}
```

# Load & Check Raw Data

## Kristiansand

```{r, message=FALSE}
kris_raw <- read_csv(str_c(path, "AlbaMotes_datakristiansand_ManyPrimates.csv"))
```

```{r}
do_checks(kris_raw)
```

```{r}
# change data file
kris <- kris_raw %>%
  filter(!is.na(pick)) %>%
  mutate(test_situation = ifelse(test_situation == "solitary", "single", "group"),
         species = fct_recode(species, "black faced spider monkey" = "black faced spider monkeys"),
         species = gsub(" ", "_", species))
```

## Kumamoto

```{r}
kuma_raw <- read_excel(str_c(path, "ManyPrimates Pilot Coding -kano.xlsx"), 2)
```

```{r}
do_checks(kuma_raw)
```

```{r}
# change data file
kuma <- kuma_raw %>%
  select(-`Second Coder`) %>%
  mutate(condition = dplyr::recode(condition, "0s" = "short", "15s" = "medium", "30s" = "long"))
```

## Edinburgh

### Capuchins

```{r}
edin_cap_raw <- read_excel("../data/raw_data/ManyPrimates Pilot_CapuchinsEdi_Voelter.xlsx")
```

```{r}
do_checks(edin_cap_raw)
```

```{r}
# change data file
edin_cap <- edin_cap_raw %>%
  mutate(researcher = "voelter") %>%
  group_by(subject_site) %>%
  mutate(trial = 1:n()) # to make trial continous across blocks
```

### Squirrel Monkeys

```{r}
edin_squ_raw <- read_excel(str_c(path, "ManyPrimates Pilot Coding_Squirrel monkeys_EW_17.7.18.xlsx"))
```

```{r}
do_checks(edin_squ_raw)
```

```{r}
# rename data file
edin_squ <- edin_squ_raw
```

### Chimps

```{r}
edin_chimps_raw <- read_excel(str_c(path, "ManyPrimates Pilot Coding_edin_chimps_prel.xlsx"))
```

```{r}
do_checks(edin_chimps_raw)
```

```{r}
# change data file
edin_chimps <- edin_chimps_raw %>%
  filter(!is.na(pick)) %>%
  mutate(cup_distance = str_sub(cup_distance, 1, -4) %>% as.numeric,
         board_size = str_sub(board_size, 1, 2) %>% as.numeric,
         researcher = "call") %>%
  group_by(subject_site) %>%
  mutate(trial = 1:n()) # to make trial continous across blocks
```

## Sweetwaters

### Chimps 1

```{r, message=FALSE}
swee_1_raw <- read_delim(str_c(path, "Data_sweetwaters.csv"), delim = ";")
```

```{r}
do_checks(swee_1_raw)
```

```{r}
# change data file
swee_1 <- swee_1_raw %>%
  filter(!is.na(pick)) %>%
  mutate(researcher = "haux",
         species = "chimpanzee")
```

### Chimps 2

```{r}
swee_2_raw <- read_excel(str_c(path, "manyprimatespilot_sweetwaters_group2.xlsx"))
```

```{r}
do_checks(swee_2_raw)
```

```{r}
# change data file
swee_2 <- swee_2_raw %>%
  filter(!is.na(pick)) %>%
  mutate(species = "chimpanzee",
         secondary_pick = as.character(secondary_pick))
```

## Deutsches Primatenzentrum / German Primate Center

### Lemurs

```{r}
dpz_lem_raw <- read_excel(str_c(path, "data_ruffed_ringtailed_lemurs_C_Fichtel.xlsx"))
```

```{r}
do_checks(dpz_lem_raw)
```

```{r}
# change data file
dpz_lem <- dpz_lem_raw %>%
  filter(pick != 9) %>%
  mutate(condition = dplyr::recode(condition, "0s" = "short", "15s" = "medium", "30s" = "long"),
         species = dplyr::recode(species,
                          "black-ad white ruffed lemur" = "black_and_white_ruffed_lemur",
                          "ring-tailed lemurs" = "ring_tailed_lemur"),
         cup_distance = str_sub(cup_distance, 1, -4) %>% as.numeric,
         board_size = str_sub(board_size, 1, 2) %>% as.numeric,
         date = gsub("-", "", date) %>% as.numeric,
         cup_distance = 17) # initially measured cup side to side, corrected measure sent via email
```

### Long-tailed Macaques

```{r}
dpz_mac_raw <- read_excel(str_c(path, "ManyPrimates_data dpz_long-tailed macaques.xlsx"))
```

```{r}
do_checks(dpz_mac_raw)
```

```{r}
# change data file
dpz_mac <- dpz_mac_raw %>%
  mutate(site = toupper(site),
         species = gsub("[ -]", "_", species),
         correct = as.numeric(correct),
         secondary_pick = as.character(secondary_pick),
         cup_distance = 16) # initially measured cup side to side, corrected measure sent via email
```

## Lincoln Park Zoo

```{r}
linc_raw <- read_excel(str_c(path, "LPZ 3 cups data June 2018 Hopper Egelkamp Ross.xlsx"))
```

```{r}
do_checks(linc_raw)
```

```{r}
# change data file
linc <- linc_raw %>%
  mutate(species = tolower(species),
         board_size = 61.5, # 35.5 was board depth; corrected via email w L Hopper
         cup_distance = 26.67) # initially measured cup side to side, corrected measure sent via email
```

## Language Research Center

```{r}
# specify column types -- makes 1st col (date) text, guesses (as usual) for everything else
ct <- c("text", rep("guess", 17))
```

```{r}
lrc_cap_raw <- read_excel(str_c(path, "ManyPrimates Pilot_LRC_ACCI.xlsx"), 6, col_types = ct)
```

```{r}
do_checks(lrc_cap_raw)
```

```{r}
# change data file

# A NOTE ON THE DATES HERE (--jw)
# In the raw file, some of the dates were in Excel date format, some in yyyymmdd format
# I initially read them in as text (because it wouldn"t let me read them as numeric)
# Here, 1) turn them all into numbers (e.g., 43262 = Excel date for 2018-06-11)
# 2) Convert numbers smaller than 20180000 (i.e., the Excel date numbers) into dates
# 3) Then remove the dashes and make numeric (e.g., "2018-06-11" --> 20180611)

lrc_cap <- lrc_cap_raw %>%
  mutate(species = "brown_capuchin_monkey",
         age = ifelse(age == "Unknown (~40)", 40, age) %>% as.numeric,
         cup_distance = 13.5, # initially measured side to side, corrected measure sent via email
         board_size = 56,
         date = as.numeric(date),
         date = ifelse(date < 20180000, as.character(as.Date(date, origin = "1899-12-30")), date),
         date = gsub("-", "", date) %>% as.numeric)
```

## Ape Cognition & Conservation Initiative

```{r}
acci_bono_raw <- read_excel(str_c(path, "ManyPrimates Pilot_LRC_ACCI.xlsx"), 7)
```

```{r}
do_checks(acci_bono_raw)
```

```{r}
# change data file
acci_bono <- acci_bono_raw %>%
  filter(!is.na(pick)) %>%
  rename(comment = X__1) %>%
  mutate(board_size = 50.8, # 20 in to cm
         cup_distance = 24) # initially measured side to side, corrected measure sent via email
```

## Leipzig

```{r}
lpz_orang_chimps_raw <- read_excel(str_c(path, "ManyPrimates Pilot_Lpz_Mkyhv_data.xlsx"), 5)
```

```{r}
do_checks(lpz_orang_chimps_raw)
```

```{r}
# change data file
lpz_orang_chimps <- lpz_orang_chimps_raw %>%
  rename(comment = X__1) %>%
  group_by(subject_site) %>%
  mutate(trial = 1:n()) # to make trial continous across blocks
```

## Monkey Haven

```{r}
mkyhv_mcqs_raw <- read_excel(str_c(path, "ManyPrimates Pilot_Lpz_Mkyhv_data.xlsx"), 8)
```

```{r}
do_checks(mkyhv_mcqs_raw)
```

```{r}
# change data file
mkyhv_mcqs <- mkyhv_mcqs_raw %>%
  mutate(pick = as.numeric(pick),
         correct = as.numeric(correct),
         species = gsub(" ", "_", species),
         test_situation = dplyr::recode(test_situation, "double" = "group"),
         secondary_pick = as.character(secondary_pick),
         cup_distance = 18) %>% # initially measured side to side, corrected measure sent via email 
  filter(!is.na(pick))
```

## Heidelberg

```{r}
hbg_raw <- read_excel(str_c(path, "Rohdaten Zoo Heidelberg.xlsx"))
```

```{r}
do_checks(hbg_raw)
```

```{r}
# change data file
hbg <- hbg_raw %>%
  filter(subject_site != "Bobo_Heidelberg") %>% # subject did not complete enough trials
  mutate(pick = as.numeric(pick),
         species = gsub(" ", "_", species),
         secondary_pick = as.character(secondary_pick)) %>%
  group_by(subject_site) %>%
  mutate(trial = 1:n()) %>% # to make trial continous across blocks
  filter(!is.na(pick))
```

# Merge Data Frames

```{r merge data files}
raw_data <- bind_rows(
  kuma,
  kris,
  edin_cap,
  edin_squ,
  swee_1,
  swee_2,
  dpz_lem,
  dpz_mac,
  lrc_cap,
  linc,
  acci_bono,
  edin_chimps,
  lpz_orang_chimps,
  mkyhv_mcqs,
  hbg) %>%
  select(-comment, -date, -secondary_pick)
```

> - Add life expectancy and normalize it
> - Make `correct` value match whether `hiding_location` and `pick` are identical

```{r add life expectancy, message=FALSE}
species_data <- read_csv("../data/species_data.csv") %>% select(species, life_expectancy)
raw_data <- raw_data %>%
  left_join(species_data, by = "species") %>%
  rename(delay = condition) %>%
  mutate(norm_age = scale(age / life_expectancy),
         correct = as.numeric(hiding_location == pick))
```

## Quality Checks

> Data structure

```{r}
glimpse(raw_data)
```

> Sample sizes

```{r, results="asis"}
raw_data %>%
  summarise(N = n_distinct(subject_site),
            n_sites = n_distinct(site) - 1, # sweetwaters and sweetwaters_group2
            n_species = n_distinct(species)) %>%
  knitr::kable()
```

```{r, results="asis"}
raw_data %>%
  group_by(site, species) %>%
  summarise(n = n_distinct(subject_site)) %>%
  knitr::kable()
```

> Are there subjects with same trial number multiple times?

```{r}
raw_data %>%
  group_by(site, species, subject_site, trial) %>%
  summarise(n = n()) %>%
  filter(n > 1)
```

> Are there any missing values?

```{r}
raw_data %>%
  select_if(function(x) any(is.na(x))) %>%
  summarise_all(funs(sum(is.na(.))))
```

> Export Merged Data

```{r write data file}
write_csv(raw_data, "../data/merged_data/01_manyprimates_pilot_merged_data_v2.csv")
```
