==> analysis/02_visualization.Rmd <==
---
title: "ManyPrimates plots"
date: "May 31 2019"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r setup, message=FALSE}
library(tidyverse)
library(ggthemes)
library(gridExtra)
library(viridis)
library(colorblindr)
library(ggpubr) # for ggarrange
library(langcog) # for multi_boot_standard

theme_set(theme_few())

# langcog package is installed by running
# install.packages("devtools")
# devtools::install_github("langcog/langcog")
```

```{r}
# make a base plots
# for flipped coordinate system with prop correct on x
p1 <- ggplot() +
  geom_hline(yintercept = 1/3, lty = 2) +
  scale_y_continuous("Proportion Correct", limits = c(-0.05, 1.05)) +
  theme(axis.title.y = element_blank(), axis.ticks.y = element_blank())

p2 <- p1 + theme(axis.text.y = element_blank())

cols = viridis(11)[c(1, 5, 9)]
cols4 = viridis(15)[c(1, 5, 9, 13)]
```

# Intro

These are some preliminary visualizations of the data.

```{r loading data, message=FALSE}
phylo <- read_csv("../data/species_data.csv") %>% 
  select(species, species_formatted, clade, clade_formatted, phylo)

mp_data <- read.csv("../data/merged_data/01_manyprimates_pilot_merged_data_v2.csv",
                    stringsAsFactors = F) %>%
  left_join(phylo, by = "species") %>%
  mutate(species = species_formatted,
         species = reorder(species, phylo),
         clade = clade_formatted,
         clade = fct_relevel(clade, "Lemur", "New World monkey", "Old World monkey", "Ape"),
         clade = fct_recode(clade, "New World Monkey" = "New World monkey", "Old World Monkey" = "Old World monkey"),
         delay = factor(delay),
         delay = fct_recode(delay, "0 seconds" = "short", "15 seconds" = "medium", "30 seconds" = "long"),
         site = fct_recode(site, 
                           "Edinburgh Zoo" = "edinburgh", 
                           "Kumamoto Sanctuary" = "kumamoto sanctuary",
                           "Wolfgang Köhler Primate\nResearch Center" = "Leipzig Zoo",
                           "Sweetwaters Chimpanzee   \nSanctuary Group 1" = "sweetwaters",
                           "Sweetwaters Chimpanzee\nSanctuary Group 2" = "sweetwaters_group2",
                           "Language Research Center" = "LRC"
                           )
)
```

# Overview by species

First an overview of the data, plotted by species and delay. Small, transparent dots represent aggregated data for each individual by delay. Open dots are the group mean for that delay. Error bars are 95% confidence intervals. Text labels are the sample size.

```{r overview by species, warning=FALSE, message=FALSE}
plot_individual <- mp_data %>%
  group_by(phylo, clade, species, delay, subject_site) %>%
  summarise(correct = mean(correct),
            nr_trials = max(trial))

plot_group <- plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r}
# get sample sizes (for use in plots)
ns <- mp_data %>%
  group_by(clade, species) %>%
  mutate(n = n_distinct(subject_site)) %>%
  group_by(clade, species, site, n) %>%
  summarise(n_by_site = n_distinct(subject_site))
```

```{r, fig.width=4, fig.height=4.5}
p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = plot_individual, aes(size = nr_trials), width = .1, height = .015, alpha = .15) +
  geom_point(data = plot_group, aes(y = mean, shape = delay), size = 3, stroke = 1.5) +
  geom_linerange(data = plot_group, aes(y = NULL, ymin = ci_lower, ymax = ci_upper), lwd = 1.2, show.legend = FALSE) +
  geom_text(data = ns, aes(label = str_c("n = ", n)), y = -.11, x = 2, col = "black", size = 3, hjust = 0) +
  facet_grid(species ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180, hjust = 0)) +
  scale_size_area("Number of Trials", max_size = 3, breaks = c(12, 24, 36)) +
  scale_shape_manual("Delay", values = c(1, 2, 5), breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  scale_color_manual("Delay", values = cols, breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  coord_flip(ylim = c(-.08, 1))
```
```{r}
ggsave("../graphs/02_01_overview.png", width = 4, height = 4.5, scale = 2)
ggsave("../graphs/Fig1.tiff", width = 4, height = 4.5, scale = 2, type = "cairo", compression = "lzw")
```

```{r, fig.width=8, fig.height=9}
# simulate color vision deficiencies
cvd_grid()
```

# Overview by species across delays

```{r overview by species across delays, warning=FALSE}
plot_individual2 <- plot_individual %>%
  group_by(clade, species, subject_site) %>%
  summarise(correct = mean(correct))

plot_group2 <- plot_individual2 %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.8}
p1 + aes(x = fct_rev(species), y = correct, col = clade) +
  geom_jitter(data = plot_individual2, width = .1, height = .015, alpha = .15, size = 3) +
  geom_point(data = plot_group2, aes(y = mean), shape = 1, size = 3, stroke = 1.5) +
  geom_linerange(data = plot_group2, aes(y = NULL, ymin = ci_lower, ymax = ci_upper), lwd = 1.2) +
  geom_text(data = ns, aes(label = n), y = -.05, col = "black", size = 3) +
  facet_grid(clade ~ ., scales = "free_y", space = "free_y") +
  theme(strip.text.y = element_blank()) + 
  scale_colour_solarized() +
  coord_flip()
```

```{r}
ggsave("../graphs/02_02_overview_across_delays.png", width = 4, height = 1.8, scale = 2)
```

# Plots by site

Here we select the species for which we have data from multiple sites. This is a very preliminary way of checking whether there is a lot of variation between sites. Plotting conventions are the same as above.

First we check for which species we have data from more than one site:

```{r plots by site, results="asis"}
mp_data %>%
  group_by(species) %>%
  summarise(sites = n_distinct(site)) %>%
  arrange(desc(sites)) %>%
  knitr::kable()
```

## Chimpanzees

```{r}
chimp_plot_individual <- mp_data %>%
  filter(species == "Chimpanzee") %>%
  group_by(site, delay, subject_site) %>%
  summarise(nr_trials = max(trial), correct = mean(correct))

chimp_plot_group <- chimp_plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=2.5}
ch <- p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = chimp_plot_individual, aes(size = nr_trials), width = .1, height = .015, alpha = .3) +
  geom_pointrange(data = chimp_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper, shape = delay), size = .8, stroke = 1.5) +
  geom_text(data = filter(ns, species == "Chimpanzee"), aes(label = str_c("n = ", n_by_site)), y = -.11, x = 2, col = "black", size = 3, hjust = 0) +
  facet_grid(site ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180, hjust = 0)) +
  scale_size_area("Number of Trials", max_size = 3, breaks = c(12, 24, 36)) +
  scale_shape_manual("Delay", values = c(1, 2, 5), breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  scale_color_manual("Delay", values = cols, breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  ggtitle("Chimpanzees") +
  coord_flip(ylim = c(-.08, 1))

ch
```

## Ring-tailed lemurs

```{r}
rtlemur_plot_individual <- mp_data %>%
  filter(species == "Ring-tailed lemur") %>%
  group_by(site, delay, subject_site) %>%
  summarise(correct = mean(correct))

rtlemur_plot_group <- rtlemur_plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = rtlemur_plot_individual, width = .1, height = .015, alpha = .3, size = 3) +
  geom_pointrange(data = rtlemur_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper), size = .8, shape = 1, stroke = 1.5) +
  geom_text(data = filter(ns, species == "Ring-tailed lemur"), aes(label = n_by_site), y = -.05, x = 2, col = "black", size = 3) +
  facet_grid(site ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180)) +
  scale_colour_solarized(breaks = c("short", "medium", "long")) +
  ggtitle("Ring-tailed lemurs") +
  coord_flip()
```

```{r}
ggsave("../graphs/02_04_rtlemur_by_site.png", width = 4, height = 1.2, scale = 2)
```


## Brown Capuchins

```{r}
cap_plot_individual <- mp_data %>%
  filter(species == "Brown capuchin monkey") %>%
  group_by(site, delay, subject_site) %>%
  summarise(nr_trials = max(trial), correct = mean(correct))

cap_plot_group <- cap_plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
cap <- p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = cap_plot_individual, aes(size = nr_trials), width = .1, height = .015, alpha = .3) +
  geom_pointrange(data = cap_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper, shape = delay), size = .8, stroke = 1.5) +
  geom_text(data = filter(ns, species == "Brown capuchin monkey"), aes(label = str_c("n = ", n_by_site)), y = -.11, x = 2, col = "black", size = 3, hjust = 0) +
  facet_grid(site ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180, hjust = 0)) +
  scale_size_area("Number of Trials", max_size = 3, breaks = c(12, 24, 36)) +
  scale_shape_manual("Delay", values = c(1, 2, 5), breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  scale_color_manual("Delay", values = cols, breaks = c("0 seconds", "15 seconds", "30 seconds")) +
  ggtitle("Capuchin Monkeys") +
  coord_flip(ylim = c(-.08, 1))
```

```{r}
fig5 <- ggarrange(ch, cap, ncol = 1, nrow = 2, heights = c(2, 1), common.legend = T, legend = "right")
```

```{r, fig.width=4, fig.height=4}
fig5
```

```{r}
ggsave("../graphs/02_05_chimp_capuchin_by_site.png", fig5, width = 4, height = 4, scale = 2)
ggsave("../graphs/FigS3.tiff", fig5, width = 4, height = 4, scale = 2, type = "cairo", compression = "lzw")
```

```{r, fig.width=8, fig.height=6}
# simulate color vision deficiencies
cvd_grid(fig5)
```

## Bonobos

```{r}
bon_plot_individual <- mp_data %>%
  filter(species == "Bonobo") %>%
  group_by(site, delay, subject_site) %>%
  summarise(correct = mean(correct))

bon_plot_group <- bon_plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = bon_plot_individual, width = .1, height = .015, alpha = .3, size = 3) +
  geom_pointrange(data = bon_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper), size = .8, shape = 1, stroke = 1.5) +
  geom_text(data = filter(ns, species == "Bonobo"), aes(label = n_by_site), y = -.05, x = 2, col = "black", size = 3) +
  facet_grid(site ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180)) +
  scale_colour_solarized(breaks = c("short", "medium", "long")) +
  ggtitle("Bonobos") +
  coord_flip()
```

```{r}
ggsave("../graphs/02_06_bonobo_by_site.png", width = 4, height = 1.2, scale = 2)
```

## Gorilla

```{r}
gor_plot_individual <- mp_data %>%
  filter(species == "Gorilla") %>%
  group_by(site, delay, subject_site) %>%
  summarise(correct = mean(correct))

gor_plot_group <- gor_plot_individual %>%
   multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
p2 + aes(x = delay, y = correct, col = delay) +
  geom_jitter(data = gor_plot_individual, width = .1, height = .015, alpha = .3, size = 3) +
  geom_pointrange(data = gor_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper), size = .8, shape = 1, stroke = 1.5) +
  geom_text(data = filter(ns, species == "Gorilla"), aes(label = n_by_site), y = -.05, x = 2, col = "black", size = 3) +
  facet_grid(site ~ ., switch = "y") +
  theme(strip.text.y = element_text(angle = 180)) +
  scale_colour_solarized(breaks = c("short", "medium", "long")) +
  ggtitle("Gorillas") +
  coord_flip()
```

```{r}
ggsave("../graphs/02_07_gorilla_by_site.png", width = 4, height = 1.2, scale = 2)
```

# Task-experience

Here we split each species by task experience. Check if we have species with sufficient memebers having different levels of task experience.

```{r plots by task experience, results="asis"}
mp_data %>%
  group_by(species) %>%
  mutate(lvls_task_exp = n_distinct(task_experience)) %>%
  filter(lvls_task_exp > 1) %>%
  group_by(species, task_experience) %>%
  summarise(n = n_distinct(subject_site)) %>%
  knitr::kable()
```

```{r}
# get sample sizes (for use in plots)
ns_task_exp <- mp_data %>%
  group_by(clade, species) %>%
  mutate(n = n_distinct(task_experience)) %>%
  group_by(clade, species, task_experience, n) %>%
  summarise(n_by_site = n_distinct(subject_site))
```

So far, this only applies to chimps and bonobos. However, task experience co-varies with site.

<!-- ## Chimpanzees -->

```{r}
chimp_task_plot_individual <- mp_data %>%
  filter(species == "Chimpanzee") %>%
  group_by(site, task_experience, delay, subject_site) %>%
  summarise(correct = mean(correct))

chimp_task_plot_group <- mp_data %>%
  filter(species == "Chimpanzee") %>%
  group_by(task_experience, delay, subject_site) %>%
  summarise(correct = mean(correct)) %>%
  multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
p_taskexp_chimp <- p2 + aes(x = delay, y = correct) +
  geom_jitter(data = chimp_task_plot_individual, aes(col = site), width = .3, height = .015, alpha = .5, size = 3) +
  geom_pointrange(data = chimp_task_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper, shape = delay), size = .8, stroke = 1.5) +
  geom_text(data = filter(ns_task_exp, species == "Chimpanzee"), aes(label = n_by_site), y = -.05, x = 2, col = "black", size = 3) +
  facet_grid(task_experience ~ ., switch = "y") +
  theme(legend.box = "horizontal", strip.text.y = element_text(angle = 180)) +
  scale_shape_manual(values = c(1, 2, 5), breaks = c("short", "medium", "long")) +
  scale_colour_solarized() +
  ggtitle("Chimpanzees") +
  coord_flip()
```

<!-- ## Bonobos -->

```{r}
bonobo_task_plot_individual <- mp_data %>%
  filter(species == "Bonobo") %>%
  group_by(site, task_experience, delay, subject_site) %>%
  summarise(correct = mean(correct))

bonobo_task_plot_group <- mp_data %>%
  filter(species == "Bonobo") %>%
  group_by(task_experience, delay, subject_site) %>%
  summarise(correct = mean(correct)) %>%
  multi_boot_standard(col = "correct")
```

```{r, fig.width=4, fig.height=1.2}
p_taskexp_bon <- p2 + aes(x = delay, y = correct) +
  geom_jitter(data = bonobo_task_plot_individual, aes(col = site), width = .3, height = .015, alpha = .5, size = 3) +
  geom_pointrange(data = bonobo_task_plot_group, aes(y = mean, ymin = ci_lower, ymax = ci_upper, shape = delay), size = .8, stroke = 1.5) +
  geom_text(data = filter(ns_task_exp, species == "Bbonobo"), aes(label = n_by_site), y = -.05, x = 2, col = "black", size = 3) +
  facet_grid(task_experience ~ ., switch = "y") +
  theme(legend.box = "horizontal", strip.text.y = element_text(angle = 180)) +
  scale_shape_manual(values = c(1, 2, 5), breaks = c("short", "medium", "long")) +
  scale_colour_solarized() +
  ggtitle("Bonobos") +
  coord_flip()
```

```{r, fig.width=4, fig.height=3}
grid.arrange(p_taskexp_chimp, p_taskexp_bon, ncol = 1)
```

```{r}
grob <- arrangeGrob(p_taskexp_chimp, p_taskexp_bon, ncol = 1)
ggsave("../graphs/02_08_task_experience.png", grob, width = 4, height = 3, scale = 2)
```


# Age

Here we plot age against correct choice separate for each delay and species. Regression line is smoothed delayal mean.

```{r}
plot_age <- mp_data %>%
  mutate(delay = fct_rev(delay)) %>%
  group_by(subject_site, norm_age, clade, species, delay) %>%
  summarise(correct = mean(correct))
```

```{r, fig.width=4, fig.height=1.3}
ggplot(plot_age, aes(x = norm_age, y = correct)) +
  geom_jitter(aes(col = clade, fill = clade, shape = clade), width = .05, height = .05, alpha = .5, size = 2.5) +
  geom_smooth(method = "glm", method.args = list(family = "binomial"), col = "black") +
  # geom_vline(xintercept = 0, lty = 2) +
  geom_hline(yintercept = 1/3, lty = 2) +
  facet_grid(~ delay) +
  labs(x = "Normed Age (relative to species longevity)", y = "Proportion Correct") +
  scale_shape_manual("Clade", values = 21:24) +
  scale_color_manual("Clade", values = cols4) +
  scale_fill_manual("Clade", values = cols4) +
  ylim(c(-.05, 1.05))
```

```{r}
ggsave("../graphs/02_09_age_by_delay.png", width = 4, height = 1.3, scale = 2)
ggsave("../graphs/FigS1.tiff", width = 4, height = 1.3, scale = 2, type = "cairo", compression = "lzw")
```

```{r, fig.width=8, fig.height=2.6}
cvd_grid()
```


```{r, fig.width=4, fig.height=3.4}
ggplot(plot_age, aes(x = norm_age, y = correct)) +
  geom_jitter(aes(fill = species), width = .05, height = .05, alpha = .5, size = 2.5, shape = 21, stroke = 0) +
  geom_smooth(aes(col = delay), method = "glm", method.args = list(family = "binomial"), show.legend = F) +
  geom_vline(xintercept = 0, lty = 2) +
  geom_hline(yintercept = 1/3, lty = 2) +
  facet_grid(delay ~ clade, scales = "free_x") +
  labs(x = "Normed Age (relative to species longevity)", y = "Proportion Correct") +
  theme(legend.position = "bottom") +
  scale_color_manual(values = rev(solarized_palette(3))) +
  ylim(c(-.05, 1.05))
```

```{r}
ggsave("../graphs/02_10_age_by_delay_species.png", width = 4, height = 3.4, scale = 2)
```


# Cup distance and board size

(sig. effect in model)

```{r}
plot_cup <- mp_data %>%
   group_by(subject_site, cup_distance, clade, species, delay) %>%
   summarise(correct = mean(correct))
```

```{r plotting cup distance}
cp <- ggplot(plot_cup, aes(x = cup_distance, y = correct)) +
  geom_jitter(aes(col = clade, fill = clade, shape = clade), width = .5, height = .05, alpha = .5, size = 2.5) +
  geom_smooth(method = "glm", method.args = list(family = "binomial"), col = "black") +
  labs(x = "Cup Distance in cm", y = "Proportion Correct") +
  geom_hline(yintercept = 1/3, lty = 2) +
  scale_shape_manual("Clade", values = 21:24) +
  scale_color_manual("Clade", values = cols4) +
  scale_fill_manual("Clade", values = cols4) +
  ylim(c(-.05, 1.05)) +
  ggtitle("A. Cup Distance")
```

```{r}
plot_board <- mp_data %>%
   group_by(subject_site, board_size, clade, species, delay) %>%
   summarise(correct = mean(correct))
```

```{r plotting board size}
bp <- ggplot(plot_board, aes(x = board_size, y = correct)) +
  geom_jitter(aes(col = clade, fill = clade, shape = clade), width = .5, height = .05, alpha = .5, size = 2.5) +
  geom_smooth(method = "glm", method.args = list(family = "binomial"), col = "black") +
  labs(x = "Board Size in cm", y = "Proportion Correct") +
  geom_hline(yintercept = 1/3, lty = 2) +
  scale_shape_manual("Clade", values = 21:24) +
  scale_color_manual("Clade", values = cols4) +
  scale_fill_manual("Clade", values = cols4) +
  ylim(c(-.05, 1.05)) +
  ggtitle("B. Board Size")
```

```{r}
pcb <- ggarrange(cp, bp, common.legend = T, legend = "right")
```

```{r, fig.width=4, fig.height=1.5}
pcb
```

```{r}
ggsave("../graphs/02_11_cup_distance_board_size.png", pcb, width = 6, height = 2.3, scale = 2)
ggsave("../graphs/FigS2.tiff", pcb, width = 6, height = 2.3, scale = 2, type = "cairo", compression = "lzw")
```

```{r, fig.width=8, fig.height=3}
cvd_grid(pcb)
```
