==> analysis/03_sites_plot.Rmd <==
---
title: "ManyPrimates Pilot Sites"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
---

```{r setup, message=FALSE}
library(ggthemes)
library(tidyverse)
library(maps)
```

```{r load coordinates, message=FALSE}
labs <- read_csv("../data/site_coordinates.csv")
world <- map("world", plot = FALSE, fill = TRUE) %>% fortify()
```

```{r map, fig.width=9, fig.height=4}
ggplot(labs, aes(x = long, y = lat)) +
  geom_polygon(data = world, aes(group = group), fill = "grey") +
  geom_point(color = "black", size = 3) +
  geom_point(aes(color = Site), size = 2) +
  coord_fixed(1.3, ylim = c(-55, 83.59961)) +
  xlab("") + ylab("") +
  theme_minimal(base_size = 16) +
  theme(panel.grid = element_blank(), axis.text = element_blank())
```

```{r}
ggsave("../graphs/03_pilot_sites.png", width = 9, height = 4, scale = 2)
```
